StudentPerformanceTracker
This Program reads data from an already existing  CSV file containing the Roll number, Name, and Marks in 9 different subjects. The program has a total of 4 options when the code is run. The four options for the user are given below:
1-create:
This creates a new line where it can take input to create a new students with roll number, name and 9 subject marks
2-cgpa:
Takes a roll number input and calculates the cgpa of the corresponding person and displays it
3-Average:
Calculates the subject average of all the nine subjects and also displays the subject with best class average and the subject with the least class average 
4-CGPA diff:
Takes a roll number as input and Calculate average cgpa of the entire class and compares it with the cgpa of the student with the given roll number